requirements.txt
asgiref==3.2.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-admin==2.0.0
django-braces==1.14.0
django-excel-response2==2.0.8
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-six==1.0.4
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.0
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Markdown==3.2
oauthlib==3.1.0
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
screen==1.0.1
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
xlwt==1.3.0
